fix: remove update-notifier as unhelpful
Currently build shows a warning when run in the Netlify CLI and `@netlify/build` is not
up2date. This can happen when the CLI is outdated, but the CLI already uses update-notifier,
so there would be two warnings. This can also happen if the CLI is latest but @netlify/build
hasn't just release a new version and it is not included in the CLI yet. In this case the user
cannot do anything about this, because the CLI has a shrinkwrap and we force the version of
@netlify/build, so updating the lockfile will not help.