fix(build-info): fix detection of Remix and Hydrogen sites (#5820)
We noticed recently that the framework detection wasn't working correctly for all Remix sites. This leads to poor DX but it also affects our reporting that informs framework usage on the platform.

This has gotten pretty complex:
- Remix sites can either use Vite or the Remix Classic Compiler
  - These use different config files, different dev and build config, but almost identical npm
    dependencies (plus `vite` in the Remix Vite case)
- Hydrogen (v2) uses Remix and, as of a few months ago, it supports Remix Vite in addition to the
  Remix Classic Compiler (it only supported this previously)
- Netlify's Remix packages were renamed in late 2023, but the Remix detection code only listed the
  old names
  - Luckily, this still happened to work because it also listed the right config file, but only for
    Remix Classic Compiler sites
  - However this means that in the Remix Vite case we were failing to detect Remix. Sites were
    categorized as Vite instead.
- The `remix` package itself was renamed to `@remix-run/dev`, `@remix-run/serve`, etc. in 2023. We
  hadn't updated this here either, which exarcerbated the above issues.

You can't actually call `test()` within a `test.each` callback. I was only doing this because I
can't use `test.for` to access the test context because we're on an old version of vitest... but
anyway, a manual table test works here.