fix: resolve integration dev path relative to buildDir
Currently, programmatic executions of `buildSite` resolve a development
integration's path (specified via `netlify.toml#integrations[].dev.path`)
relative to the current working directory.

I don't think this makes any sense, honestly. I find this undocumented
behavior unintuitive; if I specify a relative path in a `netlify.toml` I
would expect it to be resolved relative to the `netlify.toml`.

This is _technically_ a breaking change, but in practice I really doubt
any users are testing extension build hooks, it's unlikely to break
anything for the platform team (_maybe_ a few tests, but we can update
the paths in those tests if it does).

I'm tempted to remove `testOpts.cwd` here because I can't find anywhere
that we use it internally and it's unintuitive, too, but I'm leaving it
in for now as an easy escape hatch in case this change breaks any tests
Composable Platform has.