fix: return integration build plugin path in development
Currently, when running a build in `context=dev` mode for a site that
installs an extension never actually installs the extension's build
plugin, if it has one.

That's because the integration package resolver doesn't return the path
to the built package (tarball), so build never picks it up as a plugin.
This changeset fixes that issue.