feat(@netlify/build-info): add React Router 7 detection
Technically we'd want to add it to Vite's `excludedNpmDependencies`, but we actually only
want to exclude RR >=7. Many RR <7 users will be using Vite, so we don't want to break
that.

But everything works fine if we just don't do anything here, since frameworks get priority
over bundlers when there are multiple matches. (I'm not really sure why we both excluding
Vite frameworks in the Vite logic).