fix(deps): replace deprecated glob@8
Installing netlify-cli prints a deprecation warning:
https://github.com/netlify/cli/issues/7029.

I was going to upgrade `glob`, but then I remembered it's a fairly large library, so why
not use this as an opportunity to replace it? See
https://github.com/es-tooling/module-replacements/blob/main/docs/modules/glob.md.

Compare the APIs:
- https://github.com/isaacs/node-glob#readme
- https://github.com/SuperchupuDev/tinyglobby?tab=readme-ov-file#api

The only differences are option names, which I've updated.

Our uses either hardcode patterns or enforce precisely documented user patterns, so we
won't hit any of the limitations of tinyglobby.