chore: fix various CI issues (#6124)
* test: specify a type for runWithBuildAndIntrospect

* ci: always run all tests

The full test suites run in <1m, so this isn't worth the complexity and risk.

* chore(deps): bump local packages in testing package

`packages/testing` is not configured in `release-please-config.json`, so release PRs don't
bump monorepo packages in that package (and vice versa)

This led to a really confusing bug here...

We'll fix the setup in a later PR. This just fixes the immediate test errors.

---------

Co-authored-by: Mateusz Bocian <mrstork@users.noreply.github.com>