fix!: remove addons (#6193)
Currently, we conditionally attempt to retrieve information about addons
in `@netlify/config#resolveConfig`. As far as I can tell, every query to
the `listServiceInstancesForSite` (addon info) Bitballoon endpoint fails
with a 500 and the following response body:

```json
{ error: '{"message":"Missing Authentication Token"}' }
```

...this, despite verifying that e.g. the accounts and site info queries
right next to it use the exact same API token and succeed.

Unfortunately, because of the way we do error handling, when this query
fails it fails all of the surrounding queries. We could change this code
path to have separate `try/catch`es for each query, but the CLI (and
probably other consumers) currently assume this behavior, so that would
be an obtrusive change.

Addons are a thing of the past and we're fairly sure no one is using
them any longer. (The CLI is certainly not.) Any consumers of this
module who still rely on this information  and for whom this query
somehow _does_ work can instead perform this query in application code.

I started out just removing this from the `@netlify/config` module, but
decided to remove addons entirely from this repo as to not leave this
confusing vestigial concept around.