feat: update serverless-functions-api dependency and remove telemetry file handling
The telemetry implementation within serverless-functions-api is no longer a separate published
file that we need to import. The implementation is now contained within the library and exposed to
consumers via the new context.getTracer method.

The tracing implementation is off by default, and can be turned on for a single request by setting
the header 'x-nf-enable-tracing'