feat(config,build): allow dev extension config to point at directory
This reduces how intrusive the changes to the dev flow introduced in
fe1a28a are by allowing a user to specify a path to an extension
directory, rather than to a tarball. It's entirely possible to develop
an extension without any awareness that the `.ntli` directory exists,
and that's probably ideal--I think we'll need to make extension build
output more obvious before we drop a change like "point at a tarball in
an undocumented location, please."

Honestly, I fully intend to remove this functionality in a future
breaking change--this is just to make the transition easier.